added skip tests fixture
